ir+metainterp: move PtrInfo enum to majit-ir; OptContext-coupled methods become PtrInfoExt extension trait
The `PtrInfo` enum and its pure-leaf methods (constructors, query
accessors, field/item mutators, guard_opcodes, getlenbound, etc.) now
live in `majit-ir::ptr_info`. Methods that depend on
`metainterp::optimizeopt::OptContext`, `majit_gc`, or
`crate::walkvirtual::VirtualVisitor` (`make_guards`, `force_box`,
`get_known_class`, `get_known_str_length`, `get_constant_string_spec`,
`strgetitem`, `visitor_dispatch_virtual_type`,
`is_immutable_and_filled_with_constants`) stay in metainterp as the
`PtrInfoExt` extension trait — same pattern as `IntBoundMakeGuards`
and `StrPtrInfoExt`.

`force_box_impl` (private helper for `force_box`) moves out of the
impl block as a module-private free function `fn force_box_impl(self_:
&mut PtrInfo, ...)` so the trait impl exposes only the public
surface.

Callers in optimizer.rs / rewrite.rs / virtualstate.rs / vstring.rs /
earlyforce.rs / heap.rs / pure.rs / optimizeopt/mod.rs now import
`PtrInfoExt` alongside their existing `PtrInfo` imports.
