+++ majit/majit-metainterp/src/optimizeopt/heap.rs
@@ -57,6 +57,7 @@ fn sort_array_index_entries_untranslated<T>(entries: &mut [(i64, T)]) {
 
 use majit_ir::{DescrRef, OopSpecIndex, Op, OpCode, OpRef, Value, descr::descr_identity};
 
+use crate::optimizeopt::info::PtrInfoExt;
 use crate::optimizeopt::{OptContext, Optimization, OptimizationResult};
 
 /// util.py:100-128 args_dict() / args_eq(): same_box semantics — identity
